Move 'apply' mod into 'state' mod and hide 'state.parts_mut()' fn
